Scripts de validación

+++ src/Libraries/DbStudio/DbStudio.ViewModels/Solutions/Details/Files/CsvFileViewModel.cs
@@ -16,7 +16,7 @@ namespace Bau.Libraries.DbStudio.ViewModels.Solutions.Details.Files
 		/// <summary>
 		///		Carga el archivo
 		/// </summary>
-		protected override DataTable LoadFile(bool countRecords, out int totalRecords)
+		protected override DataTable LoadFile(bool countRecords, out long totalRecords)
 		{
 			return new LibCsvFiles.Controllers.CsvDataTableReader(FileParameters)
 													.Load(FileName, ActualPage, RecordsPerPage, countRecords, out totalRecords);
